Fix Linter Errors and Refactor Project
