+++ package.json
@@ -41,7 +41,7 @@
   },
   "dependencies": {
     "moment": "^2.29.1"
-},
+  },
   "jest": {
     "moduleNameMapper": {
       "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js"
